+++ app/users/dashboard/workforce/employees/[id]/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { use, useState } from 'react';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/shadcn/button';
+import { Badge } from '@/components/shadcn/badge';
 import {
   AssignManagerDialog,
   AssignRoleDialog,
